Release v4.56.55 (2017-07)

Release notes can be found at https://github.com/metasfresh/metasfresh/blob/master/ReleaseNotes.md#metasfresh-45655-2017-07

+++ metasfresh-webui-api/src/main/java/de/metas/ui/web/pattribute/ASIDescriptorFactory.java
@@ -204,7 +204,7 @@ public class ASIDescriptorFactory
 	{
 		final ASILayout.Builder layout = ASILayout.builder()
 				.setCaption(attributeSetDescriptor.getCaption())
-				.setM_AttributeSet_ID(attributeSetDescriptor.getDocumentTypeId());
+				.setM_AttributeSet_ID(attributeSetDescriptor.getDocumentTypeId().toInt());
 
 		attributeSetDescriptor.getFields()
 				.stream()
